requirements.txt
discord.py[voice]==2.0.1
asyncpg==0.26.0
psutil==5.9.2
